Move share and index inside common ShowList component
